readme fixing
