+++ utiles-pyo3/pyproject.toml
@@ -4,17 +4,35 @@ build-backend = "maturin"
 
 [project]
 name = "utiles"
-version = "0.4.0"
 description = "utiles = (utils + tiles) * rust"
 requires-python = ">=3.8"
 classifiers = [
-    "Programming Language :: Rust",
+    "Intended Audience :: Developers",
+    "License :: OSI Approved :: Apache Software License",
+    "License :: OSI Approved :: MIT License",
+    "Programming Language :: Python :: 3 :: Only",
+    "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3.8",
+    "Programming Language :: Python :: 3.9",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
     "Programming Language :: Python :: Implementation :: CPython",
     "Programming Language :: Python :: Implementation :: PyPy",
+    "Programming Language :: Python",
+    "Programming Language :: Rust",
+    "Typing :: Typed",
 ]
-authors = [{ name = "Jesse Rubin", email = "jessekrubin@gmail.com" }]
-maintainers = [{ name = "Jesse Rubin", email = "jessekrubin@gmail.com" }]
+authors = [{ name = "jesse rubin", email = "jessekrubin@gmail.com" }]
+maintainers = [{ name = "jesse rubin", email = "jessekrubin@gmail.com" }]
 dependencies = []
+dynamic = ["readme", "version"]
+
+[project.urls]
+Changelog = "https://github.com/jessekrubin/utiles/blob/main/CHANGELOG.md"
+Documentation = "https://github.com/jessekrubin/utiles/blob/main/utiles-pyo3/README.md"
+Homepage = "https://github.com/jessekrubin/utiles"
+Repository = "https://github.com/jessekrubin/utiles"
 
 [project.scripts]
 utiles = "utiles.cli:cli"
